find_multilabel_issues_per_class is now public + refactored
